chore(handlers): move text and voive to gpt module
